feat: added hover for decoded secret values
